refactor: move logic to separate hook
